Tweak return type declarations and related CI checks
