+++ .github/patch-types.php
@@ -1,5 +1,7 @@
 <?php
 
+use PHPUnit\Framework\TestCase;
+
 $mode = $argv[1] ?? 'patch';
 if ('lint' !== $mode && false === getenv('SYMFONY_PATCH_TYPE_DECLARATIONS')) {
     echo "Please define the SYMFONY_PATCH_TYPE_DECLARATIONS env var when running this script.\n";
@@ -18,17 +20,16 @@ foreach ($loader->getClassMap() as $class => $file) {
 
     switch (true) {
         case false !== strpos($file, '/src/Symfony/Component/Cache/Traits/Redis'):
+        case false !== strpos($file, '/src/Symfony/Component/Cache/Traits/Relay'):
             if (!str_ends_with($file, 'Proxy.php')) {
                 break;
             }
-            // no break;
+            continue 2;
         case false !== strpos($file, '/vendor/'):
-        case false !== strpos($file, '/src/Symfony/Bridge/Doctrine/Logger/DbalLogger.php'):
         case false !== strpos($file, '/src/Symfony/Bridge/Doctrine/Middleware/Debug/'):
         case false !== strpos($file, '/src/Symfony/Bridge/Doctrine/Tests/Fixtures/LegacyQueryMock.php'):
         case false !== strpos($file, '/src/Symfony/Bridge/PhpUnit/'):
         case false !== strpos($file, '/src/Symfony/Bundle/FrameworkBundle/Tests/Fixtures/Validation/Article.php'):
-        case false !== strpos($file, '/src/Symfony/Component/Cache/Tests/Fixtures/DriverWrapper.php'):
         case false !== strpos($file, '/src/Symfony/Component/Config/Tests/Fixtures/BadFileName.php'):
         case false !== strpos($file, '/src/Symfony/Component/Config/Tests/Fixtures/BadParent.php'):
         case false !== strpos($file, '/src/Symfony/Component/Config/Tests/Fixtures/ParseError.php'):
@@ -38,39 +39,32 @@ foreach ($loader->getClassMap() as $class => $file) {
         case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/includes/compositetype_classes.php'):
         case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/includes/intersectiontype_classes.php'):
         case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/includes/MultipleArgumentsOptionalScalarNotReallyOptional.php'):
-        case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/CheckTypeDeclarationsPass/IntersectionConstructor.php'):
-        case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/NewInInitializer.php'):
         case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/ParentNotExists.php'):
         case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/Preload/'):
         case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/Prototype/BadClasses/MissingParent.php'):
         case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/'):
-        case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/TestServiceSubscriberIntersectionWithTrait.php'):
         case false !== strpos($file, '/src/Symfony/Component/ErrorHandler/Tests/Fixtures/'):
-        case false !== strpos($file, '/src/Symfony/Component/HttpClient/Internal/'):
-        case false !== strpos($file, '/src/Symfony/Component/Form/Tests/Fixtures/Answer.php'):
-        case false !== strpos($file, '/src/Symfony/Component/Form/Tests/Fixtures/Number.php'):
-        case false !== strpos($file, '/src/Symfony/Component/Form/Tests/Fixtures/Suit.php'):
+        case false !== strpos($file, '/src/Symfony/Component/HttpClient/Internal/') && str_contains($file, 'V5'):
+        case false !== strpos($file, '/src/Symfony/Component/PropertyAccess/Tests/Fixtures/AsymmetricVisibility.php'):
         case false !== strpos($file, '/src/Symfony/Component/PropertyInfo/Tests/Fixtures/'):
-        case false !== strpos($file, '/src/Symfony/Component/PropertyInfo/Tests/Fixtures/Php81Dummy.php'):
         case false !== strpos($file, '/src/Symfony/Component/Runtime/Internal/ComposerPlugin.php'):
+        case false !== strpos($file, '/src/Symfony/Component/Security/Http/Tests/Fixtures/IsGrantedAttributeMethodsWithClosureController.php'):
+        case false !== strpos($file, '/src/Symfony/Component/Security/Http/Tests/Fixtures/IsGrantedAttributeWithClosureController.php'):
         case false !== strpos($file, '/src/Symfony/Component/Serializer/Tests/Fixtures/'):
         case false !== strpos($file, '/src/Symfony/Component/Serializer/Tests/Normalizer/Features/ObjectOuter.php'):
+        case false !== strpos($file, '/src/Symfony/Component/Validator/Tests/Constraints/Fixtures/WhenTestWithClosure.php'):
         case false !== strpos($file, '/src/Symfony/Component/Validator/Tests/Fixtures/NestedAttribute/Entity.php'):
         case false !== strpos($file, '/src/Symfony/Component/VarDumper/Tests/Fixtures/NotLoadableClass.php'):
-        case false !== strpos($file, '/src/Symfony/Component/VarDumper/Tests/Fixtures/ReflectionIntersectionTypeFixture.php'):
-        case false !== strpos($file, '/src/Symfony/Component/VarDumper/Tests/Fixtures/ReflectionUnionTypeWithIntersectionFixture.php'):
         case false !== strpos($file, '/src/Symfony/Component/VarDumper/Tests/Fixtures/VirtualProperty.php'):
         case false !== strpos($file, '/src/Symfony/Component/VarExporter/Internal'):
         case false !== strpos($file, '/src/Symfony/Component/VarExporter/Tests/Fixtures/'):
-        case false !== strpos($file, '/src/Symfony/Component/Cache/Traits/RelayProxy.php'):
-        case false !== strpos($file, '/src/Symfony/Contracts/Service/Test/ServiceLocatorTest.php'):
-        case false !== strpos($file, '/src/Symfony/Contracts/Service/Test/ServiceLocatorTestCase.php'):
+        case false !== strpos($file, '/src/Symfony/Contracts/'):
             continue 2;
     }
 
     class_exists($class);
 
-    if ('lint' !== $mode) {
+    if ('lint' !== $mode || is_subclass_of($class, TestCase::class)) {
         continue;
     }
 
@@ -78,12 +72,13 @@ foreach ($loader->getClassMap() as $class => $file) {
     foreach ($refl->getMethods() as $method) {
         if (
             $method->getReturnType()
-            || str_contains($method->getDocComment(), '@return')
+            || (str_contains($method->getDocComment(), '@return') && str_contains($method->getDocComment(), 'resource'))
             || '__construct' === $method->getName()
             || '__destruct' === $method->getName()
             || '__clone' === $method->getName()
             || $method->getDeclaringClass()->getName() !== $class
-            || str_contains($method->getDeclaringClass()->getName(), '\\Test\\')
+            || str_contains($method->getDeclaringClass()->getName(), '\\Tests\\')
+            || str_contains($method->getDeclaringClass()->getName(), '\\Test\\') && str_starts_with($method->getName(), 'test')
         ) {
             continue;
         }
@@ -95,6 +90,7 @@ foreach ($loader->getClassMap() as $class => $file) {
 if ($missingReturnTypes) {
     echo \count($missingReturnTypes)." missing return types on interfaces\n\n";
     echo implode("\n", $missingReturnTypes);
+    echo "\n";
     exit(1);
 }
