+++ src/Symfony/Component/VarDumper/Test/VarDumperTestTrait.php
@@ -45,11 +45,17 @@ trait VarDumperTestTrait
         $this->varDumperConfig['flags'] = null;
     }
 
+    /**
+     * @return void
+     */
     public function assertDumpEquals(mixed $expected, mixed $data, int $filter = 0, string $message = '')
     {
         $this->assertSame($this->prepareExpectation($expected, $filter), $this->getDump($data, null, $filter), $message);
     }
 
+    /**
+     * @return void
+     */
     public function assertDumpMatchesFormat(mixed $expected, mixed $data, int $filter = 0, string $message = '')
     {
         $this->assertStringMatchesFormat($this->prepareExpectation($expected, $filter), $this->getDump($data, null, $filter), $message);
